fix: 删除外层dom
