+++ src/Line/demos/basic.tsx
@@ -75,7 +75,7 @@ const Demo: FC<DemoProps> = props => {
     };
   };
   return (
-    <div className={styles.demo}>
+    <div className={styles.lineDemo}>
       <div className={styles.card}>
         <div className={styles.cardHeader}>
           <div className={styles.title}>卡片标题</div>
