+++ src/Line/demos/index.less
@@ -1,4 +1,4 @@
-.demo {
+.lineDemo {
   width: 100%;
   height: 5rem;
   .space {
