+++ src/Line/index.less
@@ -1,12 +1,11 @@
-@basicPrefixCls: alita-linebasic;
 @prefixCls: alita-line;
 
-.@{prefixCls} {
-  width: 100%;
-  height: 100%;
-}
+// .@{prefixCls} {
+//   width: 100%;
+//   height: 100%;
+// }
 
-.@{basicPrefixCls} {
+.@{prefixCls} {
   position: relative;
   width: 100%;
   height: 100%;
