+++ src/index.ts
@@ -1,6 +1,6 @@
 // 此文件在构建时会自动更新，请勿手动修改，详见 package.json 中的 entry script
 export { default as Column } from './Column';
 export { default as Donut } from './Donut';
-export { default as Line } from './Line';
+export { default as Line } from './Line/index';
 export { default as MultipleProgress } from './MultipleProgress';
 export { default as SingleProgress } from './SingleProgress';
